requirements.txt
streamlit==1.37.1
google-generativeai==0.7.2
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
